# swagger/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "REST interface for the Insights Platform Host Inventory application.",
    "version": "1.0.0",
    "title": "Insights Host Inventory REST Interface"
  },
  "paths": {
    "/hosts": {
      "get": {
        "operationId": "api.host.get_host_list",
        "tags": [
          "hosts"
        ],
        "summary": "Read the entire list of hosts",
        "description": "Read the entire list of all hosts available to the account/org_id. <br /><br /> Required permissions: inventory:hosts:read",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/displayName"
          },
          {
            "$ref": "#/components/parameters/fqdn"
          },
          {
            "$ref": "#/components/parameters/hostnameOrId"
          },
          {
            "$ref": "#/components/parameters/insightsId"
          },
          {
            "$ref": "#/components/parameters/providerId"
          },
          {
            "$ref": "#/components/parameters/providerType"
          },
          {
            "$ref": "#/components/parameters/branchId"
          },
          {
            "$ref": "#/components/parameters/perPageParam"
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/orderByParam"
          },
          {
            "$ref": "#/components/parameters/orderHowParam"
          },
          {
            "$ref": "#/components/parameters/stalenessParam"
          },
          {
            "$ref": "#/components/parameters/tagsParam"
          },
          {
            "$ref": "#/components/parameters/registered_with"
          },
          {
            "$ref": "#/components/parameters/filter_param"
          },
          {
            "$ref": "#/components/parameters/fields_param"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully read the hosts list.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostQueryOutput"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "api.host.delete_host_list",
        "tags": [
          "hosts"
        ],
        "summary": "Delete the entire list of hosts filtered by the given parameters",
        "description": "Delete the entire list of hosts filtered by the given parameters. <br /><br /> Required permissions: inventory:hosts:write",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/displayName"
          },
          {
            "$ref": "#/components/parameters/fqdn"
          },
          {
            "$ref": "#/components/parameters/hostnameOrId"
          },
          {
            "$ref": "#/components/parameters/insightsId"
          },
          {
            "$ref": "#/components/parameters/providerId"
          },
          {
            "$ref": "#/components/parameters/providerType"
          },
          {
            "$ref": "#/components/parameters/registered_with"
          },
          {
            "$ref": "#/components/parameters/stalenessNoDefaultsParam"
          },
          {
            "$ref": "#/components/parameters/tagsParam"
          },
          {
            "$ref": "#/components/parameters/filter_param"
          }
        ],
        "responses": {
          "202": {
            "description": "Request for deletion of filtered hosts has been accepted."
          },
          "400": {
            "description": "Invalid request."
          }
        }
      }
    },
    "/hosts/all": {
      "delete": {
        "operationId": "api.host.delete_all_hosts",
        "tags": [
          "hosts"
        ],
        "summary": "Delete all hosts on the account/org_id",
        "description": "Delete all hosts on the account/org_id.  The request must include \"confirm_delete_all=true\". <br /><br /> Required permissions: inventory:hosts:write",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/confirmDeleteAll"
          }
        ],
        "responses": {
          "202": {
            "description": "Request for deleting all hosts has been accepted."
          },
          "400": {
            "description": "Invalid request."
          }
        }
      }
    },
    "/hosts/checkin": {
      "post": {
        "operationId": "api.host.host_checkin",
        "tags": [
          "hosts"
        ],
        "summary": "Update staleness timestamps for a host matching the provided facts",
        "description": "Finds a host and updates its staleness timestamps. It uses the supplied canonical facts to determine which host to update. By default, the staleness timestamp is set to 1 hour from when the request is received; however, this can be overridden by supplying the interval. <br /><br /> Required permissions: inventory:hosts:write",
        "security": [
          {
            "ApiKeyAuth": []
          },
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Data required to create a check-in record for a host.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCheckIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully checked in Host.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateHostOut"
                }
              }
            }
          },
          "404": {
            "description": "Not Found."
          }
        }
      }
    },
    "/hosts/{host_id_list}": {
      "get": {
        "tags": [
          "hosts"
        ],
        "summary": "Find hosts by their IDs",
        "description": "Find one or more hosts by their ID. <br /><br /> Required permissions: inventory:hosts:read",
        "operationId": "api.host.get_host_by_id",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/hostIdList"
          },
          {
            "$ref": "#/components/parameters/branchId"
          },
          {
            "$ref": "#/components/parameters/perPageParam"
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/orderByParam"
          },
          {
            "$ref": "#/components/parameters/orderHowParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully searched for hosts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HostQueryOutput"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "404": {
            "description": "Host not found."
          }
        }
      },
      "delete": {
        "tags": [
          "hosts"
        ],
        "summary": "Delete hosts by IDs",
        "description": "Delete hosts by IDs <br /><br /> Required permissions: inventory:hosts:write",
        "operationId": "api.host.delete_by_id",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/hostIdList"
          },
          {
            "$ref": "#/components/parameters/branchId"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted hosts."
          },
          "400": {
            "description": "Invalid request."
          },
          "404": {
            "description": "Host not found."
          }
        }
      },
      "patch": {
        "tags": [
          "hosts"
        ],
        "summary": "Update a host",
        "description": "Update a host <br /><br /> Required permissions: inventory:hosts:write",
        "operationId": "api.host.patch_by_id",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/hostIdList"
          },
          {
            "$ref": "#/components/parameters/branchId"
          }
        ],
        "requestBody": {
          "description": "A group of fields to be updated on the host",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchHostIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated the host."
          },
          "400": {
            "description": "Invalid request."
          },
          "404": {
            "description": "Host not found."
          }
        }
      }
    },
    "/hosts/{host_id_list}/facts/{namespace}": {
      "patch": {
        "tags": [
          "hosts"
        ],
        "summary": "Merge facts under a namespace",
        "description": "Merge one or multiple hosts facts under a namespace. <br /><br /> Required permissions: inventory:hosts:write",
        "operationId": "api.host.merge_facts",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/hostIdList"
          },
          {
            "$ref": "#/components/parameters/factsNamespace"
          },
          {
            "$ref": "#/components/parameters/branchId"
          }
        ],
        "requestBody": {
          "description": "A dictionary with the new facts to merge with the original ones.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Facts"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully merged facts."
          },
          "400": {
            "description": "Invalid request."
          },
          "404": {
            "description": "Host or namespace not found."
          }
        }
      },
      "put": {
        "tags": [
          "hosts"
        ],
        "summary": "Replace facts under a namespace",
        "description": "Replace facts under a namespace <br /><br /> Required permissions: inventory:hosts:write",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "operationId": "api.host.replace_facts",
        "parameters": [
          {
            "$ref": "#/components/parameters/hostIdList"
          },
          {
            "$ref": "#/components/parameters/factsNamespace"
          },
          {
            "$ref": "#/components/parameters/branchId"
          }
        ],
        "requestBody": {
          "description": "A dictionary with the new facts to replace the original ones.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Facts"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully replaced facts."
          },
          "400": {
            "description": "Invalid request."
          },
          "404": {
            "description": "Host or namespace not found."
          }
        }
      }
    },
    "/hosts/{host_id_list}/system_profile": {
      "get": {
        "tags": [
          "hosts"
        ],
        "summary": "Return one or more hosts system profile",
        "description": "Find one or more hosts by their ID and return the id and system profile <br /><br /> Required permissions: inventory:hosts:read",
        "operationId": "api.host.get_host_system_profile_by_id",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/hostIdList"
          },
          {
            "$ref": "#/components/parameters/perPageParam"
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/orderByParam"
          },
          {
            "$ref": "#/components/parameters/orderHowParam"
          },
          {
            "$ref": "#/components/parameters/branchId"
          },
          {
            "$ref": "#/components/parameters/fields_param"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully searched for hosts.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SystemProfileByHostOut"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          },
          "404": {
            "description": "Host not found."
          }
        }
      }
    },
    "/hosts/{host_id_list}/tags": {
      "get": {
        "tags": [
          "hosts"
        ],
        "summary": "Get the tags on a host",
        "description": "Get the tags on a host <br /><br /> Required permissions: inventory:hosts:read",
        "operationId": "api.host.get_host_tags",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/hostIdList"
          },
          {
            "$ref": "#/components/parameters/perPageParam"
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/orderByParam"
          },
          {
            "$ref": "#/components/parameters/orderHowParam"
          },
          {
            "$ref": "#/components/parameters/searchParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found tags.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagsOut"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          }
        }
      }
    },
    "/hosts/{host_id_list}/tags/count": {
      "get": {
        "tags": [
          "hosts"
        ],
        "summary": "Get the number of tags on a host",
        "description": "Get the number of tags on a host <br /><br /> Required permissions: inventory:hosts:read",
        "operationId": "api.host.get_host_tag_count",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/hostIdList"
          },
          {
            "$ref": "#/components/parameters/perPageParam"
          },
          {
            "$ref": "#/components/parameters/pageParam"
          },
          {
            "$ref": "#/components/parameters/orderByParam"
          },
          {
            "$ref": "#/components/parameters/orderHowParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found tag count.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagCountOut"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "x-bearerInfoFunc": "app.auth.bearer_token_handler"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-rh-identity",
        "description": "Base64-encoded JSON identity header provided by 3Scale. Contains an account number of the user issuing the request. Format of the JSON: {\"identity\": {\"account_number\": \"12345678\"}}",
        "x-apikeyInfoFunc": "app.auth.authentication_header_handler"
      }
    },
    "parameters": {
...
      "displayName": {
        "in": "query",
        "name": "display_name",
        "schema": {
          "type": "string"
        },
        "description": "Filter by display_name",
        "required": false
      },
      "fqdn": {
        "in": "query",
        "name": "fqdn",
        "schema": {
          "type": "string"
        },
        "description": "Filter by FQDN",
        "required": false
      },
      "hostnameOrId": {
        "in": "query",
        "name": "hostname_or_id",
        "schema": {
          "type": "string"
        },
        "description": "Filter by display_name, fqdn, id",
        "required": false
      },
      "insightsId": {
        "in": "query",
        "name": "insights_id",
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Filter by insights_id",
        "required": false
      },
      "providerId": {
        "in": "query",
        "name": "provider_id",
        "schema": {
          "type": "string"
        },
        "description": "Filter by provider_id",
        "required": false
      },
      "providerType": {
        "in": "query",
        "name": "provider_type",
        "schema": {
          "type": "string",
          "enum": [
            "alibaba",
            "aws",
            "azure",
            "gcp",
            "ibm"
          ]
        },
        "description": "Filter by provider_type",
        "required": false
      },
...
    },
...
  }
}